feat(kyc): refactor KYC block to display risk labels using badges
- Add badge component to show risk labels with descriptive statuses
- Remove the KycDecision component as it is no longer needed
- Update type definitions and adjust data handling in useKycBlock
