+++ apps/backoffice-v2/src/lib/blocks/create-blocks-typed/types.ts
@@ -145,11 +145,6 @@ export type TDetailsCell = {
   isDocumentsV2: boolean;
 };
 
-export type TKycDecision = Omit<TDetailsCell, 'value' | 'type'> & {
-  type: 'kycDecision';
-  value: TDetailsCell['value'] & { riskLabels: string[] };
-};
-
 export type TNestedDetailsCell = {
   type: 'nestedDetails';
   id?: string;
@@ -287,5 +282,4 @@ export type TCell =
   | TPDFViewerCell
   | TReadOnlyDetailsCell
   | TImageCell
-  | TEditableDetailsV2Cell
-  | TKycDecision;
+  | TEditableDetailsV2Cell;
